bdkffi

Module bitcoin

Structs

Address

A bitcoin address

Amount

The Amount type can be used to express Bitcoin amounts that support arithmetic and conversion to various denominations. The operations that Amount implements will panic when overflow or underflow occurs. Also note that since the internal representation of amounts is unsigned, subtracting below zero is considered an underflow and will cause a panic.

Block

Bitcoin block. A collection of transactions with an attached proof of work.

BlockHash

A bitcoin Block hash

ControlBlock

DescriptorId

A collision-proof unique identifier for a descriptor.

FeeRate

Represents fee rate.

FinalizedPsbtResult

HashableOutPoint

An OutPoint used as a key in a hash map.

Header

Bitcoin block header. Contains all the block’s information except the actual transactions, but including a root of a merkle tree committing to all transactions in the block.

Input

A key-value map for an input of the corresponding index in the unsigned transaction.

Key

KeySource

LeafNode

Store information about taproot leaf node.

NodeInfo

Represents the node information in taproot tree. In contrast to TapTree, this is allowed to have hidden leaves as children.

OutPoint

A reference to an unspent output by TXID and output index.

Output

A key-value map for an output of the corresponding index in the unsigned transaction.

ProprietaryKey

Psbt

A Partially Signed Transaction.

Script

A bitcoin script: https://en.bitcoin.it/wiki/Script

TapKeyOrigin

TapScriptEntry

TapScriptSigKey

TapTree

Taproot Tree representing a complete binary tree without any hidden nodes.

Transaction

Bitcoin transaction. An authenticated movement of coins.

TxIn

A transcation input.

TxMerkleNode

The merkle root of the merkle tree corresponding to a block’s transactions.

TxOut

Bitcoin transaction output.

Txid

A bitcoin transaction identifier

WitnessProgram

The version and program of a Segwit address.

Wtxid

A bitcoin transaction identifier, including witness data. For transactions with no SegWit inputs, the txid will be equivalent to wtxid.

Enums

AddressData

The type of address.

ChildNumber

A child number in a derivation path

Type Aliases

DescriptorType

Network