bdkffi::bitcoin
Struct Psbt

pub struct Psbt(/* private fields */);

A Partially Signed Transaction.

Implementations

impl Psbt

pub fn new(psbt_base64: String) -> Result<Self, PsbtParseError>

Creates a new Psbt instance from a base64-encoded string.

pub fn from_unsigned_tx(tx: Arc<Transaction>) -> Result<Arc<Psbt>, PsbtError>

Creates a PSBT from an unsigned transaction.

Errors

If transactions is not unsigned.

pub fn from_file(path: String) -> Result<Self, PsbtError>

Create a new Psbt from a .psbt file.

pub fn serialize(&self) -> String

Serialize the PSBT into a base64-encoded string.

pub fn extract_tx(&self) -> Result<Arc<Transaction>, ExtractTxError>

Extracts the Transaction from a Psbt by filling in the available signature information.

Errors

ExtractTxError variants will contain either the Psbt itself or the Transaction that was extracted. These can be extracted from the Errors in order to recover. See the error documentation for info on the variants. In general, it covers large fees.

pub fn fee(&self) -> Result<u64, PsbtError>

Calculates transaction fee.

‘Fee’ being the amount that will be paid for mining a transaction with the current inputs and outputs i.e., the difference in value of the total inputs and the total outputs.

Errors

MissingUtxo when UTXO information for any input is not present or is invalid.

NegativeFee if calculated value is negative.

FeeOverflow if an integer overflow occurs.

pub fn combine(&self, other: Arc<Psbt>) -> Result<Arc<Psbt>, PsbtError>

Combines this Psbt with other PSBT as described by BIP 174.

In accordance with BIP 174 this function is commutative i.e., A.combine(B) == B.combine(A)

pub fn finalize(&self) -> FinalizedPsbtResult

Finalizes the current PSBT and produces a result indicating

whether the finalization was successful or not.

pub fn write_to_file(&self, path: String) -> Result<(), PsbtError>

Write the Psbt to a file. Note that the file must not yet exist.

pub fn json_serialize(&self) -> String

Serializes the PSBT into a JSON string representation.

pub fn spend_utxo(&self, input_index: u64) -> String

Returns the spending utxo for this PSBT’s input at input_index.

pub fn input(&self) -> Vec<Input>

The corresponding key-value map for each input in the unsigned transaction.

pub fn output(&self) -> Vec<Output>

The corresponding key-value map for each output in the unsigned transaction.