bdkffi::descriptor

Struct Descriptor

pub struct Descriptor {
    pub extended_descriptor: ExtendedDescriptor,
    pub key_map: KeyMap,
}

An expression of how to derive output scripts: https://github.com/bitcoin/bitcoin/blob/master/doc/descriptors.md

Fields
extended_descriptor: ExtendedDescriptorkey_map: KeyMap
Implementations

impl Descriptor

pub fn new( descriptor: String, network_kind: NetworkKind, ) -> Result<Self, DescriptorError>

Parse a string as a descriptor for the given network.

pub fn new_bip44( secret_key: &DescriptorSecretKey, keychain_kind: KeychainKind, network_kind: NetworkKind, ) -> Self

Multi-account hierarchy descriptor: https://github.com/bitcoin/bips/blob/master/bip-0044.mediawiki

pub fn new_bip44_public( public_key: &DescriptorPublicKey, fingerprint: String, keychain_kind: KeychainKind, network_kind: NetworkKind, ) -> Result<Self, DescriptorError>

Multi-account hierarchy descriptor: https://github.com/bitcoin/bips/blob/master/bip-0044.mediawiki

pub fn new_bip49( secret_key: &DescriptorSecretKey, keychain_kind: KeychainKind, network_kind: NetworkKind, ) -> Self

P2SH nested P2WSH descriptor: https://github.com/bitcoin/bips/blob/master/bip-0049.mediawiki

pub fn new_bip49_public( public_key: &DescriptorPublicKey, fingerprint: String, keychain_kind: KeychainKind, network_kind: NetworkKind, ) -> Result<Self, DescriptorError>

P2SH nested P2WSH descriptor: https://github.com/bitcoin/bips/blob/master/bip-0049.mediawiki

pub fn new_bip84( secret_key: &DescriptorSecretKey, keychain_kind: KeychainKind, network_kind: NetworkKind, ) -> Self

Pay to witness PKH descriptor: https://github.com/bitcoin/bips/blob/master/bip-0084.mediawiki

pub fn new_bip84_public( public_key: &DescriptorPublicKey, fingerprint: String, keychain_kind: KeychainKind, network_kind: NetworkKind, ) -> Result<Self, DescriptorError>

Pay to witness PKH descriptor: https://github.com/bitcoin/bips/blob/master/bip-0084.mediawiki

pub fn new_bip86( secret_key: &DescriptorSecretKey, keychain_kind: KeychainKind, network_kind: NetworkKind, ) -> Self

Single key P2TR descriptor: https://github.com/bitcoin/bips/blob/master/bip-0086.mediawiki

pub fn new_bip86_public( public_key: &DescriptorPublicKey, fingerprint: String, keychain_kind: KeychainKind, network_kind: NetworkKind, ) -> Result<Self, DescriptorError>

Single key P2TR descriptor: https://github.com/bitcoin/bips/blob/master/bip-0086.mediawiki

pub fn new_wsh_sortedmulti( k: u32, pks: Vec<String>, ) -> Result<Self, DescriptorError>

Create a new wsh sorted multi descriptor Errors when miniscript exceeds resource limits under p2sh context

pub fn new_sh_sortedmulti( k: u32, pks: Vec<String>, ) -> Result<Self, DescriptorError>

Create a new sh sortedmulti descriptor with threshold k and Vec of pks. Errors when miniscript exceeds resource limits under p2sh context

pub fn new_sh_wsh_sortedmulti( k: u32, pks: Vec<String>, ) -> Result<Self, DescriptorError>

Create a new sh wrapped wsh sortedmulti descriptor from threshold k and Vec of pks Errors when miniscript exceeds resource limits under segwit context

pub fn new_pk(pk: String) -> Result<Self, DescriptorError>

Create a new pay-to-pubkey descriptor from a public key string.

pub fn new_pkh(pk: String) -> Result<Self, DescriptorError>

Create a new PkH descriptor

pub fn new_wpkh(pk: String) -> Result<Self, DescriptorError>

Create a new Wpkh descriptor Will return Err if uncompressed key is used

pub fn new_sh_wpkh(pk: String) -> Result<Self, DescriptorError>

Create a new sh wrapped wpkh from Pk. Errors when uncompressed keys are supplied

pub fn new_wsh(mini_script: String) -> Result<Self, DescriptorError>

Create a new wsh descriptor from witness script Errors when miniscript exceeds resource limits under p2sh context or does not type check at the top level

pub fn new_sh_wsh(mini_script: String) -> Result<Self, DescriptorError>

Create a new sh wrapped wsh descriptor with witness script Errors when miniscript exceeds resource limits under wsh context or does not type check at the top level

pub fn new_sh(mini_script: String) -> Result<Self, DescriptorError>

Create a new sh for a given redeem script Errors when miniscript exceeds resource limits under p2sh context or does not type check at the top level

pub fn new_bare(mini_script: String) -> Result<Self, DescriptorError>

Create a new bare descriptor from witness script Errors when miniscript exceeds resource limits under bare context or does not type check at the top level

pub fn to_string_with_secret(&self) -> String

Dangerously convert the descriptor to a string.

pub fn is_multipath(&self) -> bool

Does this descriptor contain paths: https://github.com/bitcoin/bips/blob/master/bip-0389.mediawiki

pub fn descriptor_id(&self) -> Arc<DescriptorId>

A unique identifier for the descriptor.

pub fn to_single_descriptors( &self, ) -> Result<Vec<Arc<Descriptor>>, MiniscriptError>

Return descriptors for all valid paths.

pub fn max_weight_to_satisfy(&self) -> Result<u64, DescriptorError>

Computes an upper bound on the difference between a non-satisfied TxIn’s segwit_weight and a satisfied TxIn’s segwit_weight.

pub fn desc_type(&self) -> DescriptorType

pub fn derive_address( &self, index: u32, network: Network, ) -> Result<Arc<Address>, DescriptorError>

pub fn has_wildcard(&self) -> bool

Whether or not the descriptor has any wildcards.

pub fn sanity_check(&self) -> Result<(), DescriptorError>

Checks whether the descriptor is safe.

Checks whether all the spend paths in the descriptor are possible on the bitcoin network under the current standardness and consensus rules. Also checks whether the descriptor requires signatures on all spend paths and whether the script is malleable.

In general, all the guarantees of miniscript hold only for safe scripts. The signer may not be able to find satisfactions even if one exists.