bdkffi::electrum
Struct ElectrumClient

pub struct ElectrumClient(/* private fields */);

Wrapper around an electrum_client::ElectrumApi which includes an internal in-memory transaction cache to avoid re-fetching already downloaded transactions.

Implementations

impl ElectrumClient

pub fn new(url: String, socks5: Option<String>) -> Result<Self, ElectrumError>

Creates a new bdk client from a electrum_client::ElectrumApi Optional: Set the proxy of the builder

pub fn full_scan( &self, request: Arc<FullScanRequest>, stop_gap: u64, batch_size: u64, fetch_prev_txouts: bool, ) -> Result<Arc<Update>, ElectrumError>

Full scan the keychain scripts specified with the blockchain (via an Electrum client) and returns updates for bdk_chain data structures.

request: struct with data required to perform a spk-based blockchain client full scan, see FullScanRequest.

stop_gap: the full scan for each keychain stops after a gap of script pubkeys with no associated transactions.

batch_size: specifies the max number of script pubkeys to request for in a single batch request.

fetch_prev_txouts: specifies whether we want previous TxOuts for fee calculation. Note that this requires additional calls to the Electrum server, but is necessary for calculating the fee on a transaction if your wallet does not own the inputs. Methods like Wallet.calculate_fee and Wallet.calculate_fee_rate will return a CalculateFeeError::MissingTxOut error if those TxOuts are not present in the transaction graph.

pub fn sync( &self, request: Arc<SyncRequest>, batch_size: u64, fetch_prev_txouts: bool, ) -> Result<Arc<Update>, ElectrumError>

Sync a set of scripts with the blockchain (via an Electrum client) for the data specified and returns updates for bdk_chain data structures.

request: struct with data required to perform a spk-based blockchain client sync, see SyncRequest.

batch_size: specifies the max number of script pubkeys to request for in a single batch request.

fetch_prev_txouts: specifies whether we want previous TxOuts for fee calculation. Note that this requires additional calls to the Electrum server, but is necessary for calculating the fee on a transaction if your wallet does not own the inputs. Methods like Wallet.calculate_fee and Wallet.calculate_fee_rate will return a CalculateFeeError::MissingTxOut error if those TxOuts are not present in the transaction graph.

If the scripts to sync are unknown, such as when restoring or importing a keychain that may include scripts that have been used, use full_scan with the keychain.

pub fn transaction_broadcast( &self, tx: &Transaction, ) -> Result<Arc<Txid>, ElectrumError>

Broadcasts a transaction to the network.

pub fn fetch_tx( &self, txid: Arc<Txid>, ) -> Result<Arc<Transaction>, ElectrumError>

Fetch transaction of given Txid.

If it hits the cache it will return the cached version and avoid making the request.

pub fn server_features(&self) -> Result<ServerFeaturesRes, ElectrumError>

Returns the capabilities of the server.

pub fn estimate_fee(&self, number: u64) -> Result<f64, ElectrumError>

Estimates the fee required in bitcoin per kilobyte to confirm a transaction in number blocks.

pub fn block_header(&self, height: u64) -> Result<Header, ElectrumError>

Gets the block header for height height.

pub fn block_headers_subscribe( &self, ) -> Result<HeaderNotification, ElectrumError>

Subscribes to notifications for new block headers, by sending a blockchain.headers.subscribe call.

pub fn block_headers_pop( &self, ) -> Result<Option<HeaderNotification>, ElectrumError>

Tries to pop one queued notification for a new block header that we might have received. Returns None if there are no items in the queue.

pub fn ping(&self) -> Result<(), ElectrumError>

Pings the server.

pub fn relay_fee(&self) -> Result<f64, ElectrumError>

Returns the minimum accepted fee by the server’s node in Bitcoin, not Satoshi.

pub fn transaction_get_raw( &self, txid: Arc<Txid>, ) -> Result<Vec<u8>, ElectrumError>

Gets the raw bytes of a transaction with txid. Returns an error if not found.