bdkffi::esplora
Struct EsploraClient

pub struct EsploraClient(/* private fields */);

Wrapper around an esplora_client::BlockingClient which includes an internal in-memory transaction cache to avoid re-fetching already downloaded transactions.

Implementations

impl EsploraClient

pub fn new(url: String, proxy: Option<String>) -> Self

Creates a new bdk client from an esplora_client::BlockingClient. Optional: Set the proxy of the builder.

pub fn full_scan( &self, request: Arc<FullScanRequest>, stop_gap: u64, parallel_requests: u64, ) -> Result<Arc<Update>, EsploraError>

Scan keychain scripts for transactions against Esplora, returning an update that can be applied to the receiving structures.

request provides the data required to perform a script-pubkey-based full scan (see FullScanRequest). The full scan for each keychain (K) stops after a gap of stop_gap script pubkeys with no associated transactions. parallel_requests specifies the maximum number of HTTP requests to make in parallel.

pub fn sync( &self, request: Arc<SyncRequest>, parallel_requests: u64, ) -> Result<Arc<Update>, EsploraError>

Sync a set of scripts, txids, and/or outpoints against Esplora.

request provides the data required to perform a script-pubkey-based sync (see SyncRequest). parallel_requests specifies the maximum number of HTTP requests to make in parallel.

pub fn broadcast(&self, transaction: &Transaction) -> Result<(), EsploraError>

Broadcast a Transaction to Esplora.

pub fn get_tx( &self, txid: Arc<Txid>, ) -> Result<Option<Arc<Transaction>>, EsploraError>

Get a Transaction option given its Txid.

pub fn get_tx_no_opt( &self, txid: Arc<Txid>, ) -> Result<Arc<Transaction>, EsploraError>

Get a Transaction given its Txid.

pub fn get_height(&self) -> Result<u32, EsploraError>

Get the height of the current blockchain tip.

pub fn get_tip_hash(&self) -> Result<Arc<BlockHash>, EsploraError>

Get the BlockHash of the current blockchain tip.

pub fn get_fee_estimates(&self) -> Result<HashMap<u16, f64>, EsploraError>

Get a map where the key is the confirmation target (in number of blocks) and the value is the estimated feerate (in sat/vB).

pub fn get_block_hash( &self, block_height: u32, ) -> Result<Arc<BlockHash>, EsploraError>

Get the BlockHash of a specific block height.

pub fn get_block_by_hash( &self, block_hash: Arc<BlockHash>, ) -> Result<Option<Block>, EsploraError>

Get a Block given a particular BlockHash.

pub fn get_txid_at_block_index( &self, block_hash: Arc<BlockHash>, index: u64, ) -> Result<Option<Arc<Txid>>, EsploraError>

Get a Txid of a transaction given its index in a block with a given hash.

pub fn get_header_by_hash( &self, block_hash: Arc<BlockHash>, ) -> Result<Header, EsploraError>

Get a Header given a particular block hash.

pub fn get_tx_status(&self, txid: Arc<Txid>) -> Result<TxStatus, EsploraError>

Get the status of a Transaction given its Txid.

pub fn get_tx_info(&self, txid: Arc<Txid>) -> Result<Option<Tx>, EsploraError>

Get transaction info given its Txid.

pub fn get_address_txs( &self, address: Arc<Address>, last_seen: Option<Arc<Txid>>, ) -> Result<Vec<Tx>, EsploraError>

Get transaction history for the specified address, sorted with newest first.

Returns up to 50 mempool transactions plus the first 25 confirmed transactions. More can be requested by specifying the last txid seen by the previous query.

pub fn get_merkle_proof( &self, txid: &Txid, ) -> Result<Option<MerkleProof>, EsploraError>

Get a merkle inclusion proof for a Transaction with the given Txid.

pub fn get_output_status( &self, txid: Arc<Txid>, vout: u64, ) -> Result<Option<OutputStatus>, EsploraError>

Get the spending status of an output given a Txid and the output index.