bdkffi::keys
Struct DerivationPath

pub struct DerivationPath(/* private fields */);

A BIP-32 derivation path.

Implementations

impl DerivationPath

pub fn new(path: String) -> Result<Self, Bip32Error>

Parse a string as a BIP-32 derivation path.

pub fn master() -> Arc<Self>

Returns derivation path for a master key (i.e. empty derivation path)

pub fn is_master(&self) -> bool

Returns whether derivation path represents master key (i.e. it’s length is empty). True for m path.

pub fn len(&self) -> u64

Returns length of the derivation path

pub fn is_empty(&self) -> bool

Returns true if the derivation path is empty

pub fn child(&self, child_number: ChildNumber) -> Result<Arc<Self>, Bip32Error>

Create a new DerivationPath that is a child of this one.

pub fn extend(&self, other: &DerivationPath) -> Arc<Self>

Concatenate self with path and return the resulting new path.

pub fn to_u32_vec(&self) -> Vec<u32>

Returns the derivation path as a vector of u32 integers. Unhardened elements are copied as is. 0x80000000 is added to the hardened elements.