bdkffi::kyoto
Struct CbfBuilder

pub struct CbfBuilder { /* private fields */ }

Build a BIP 157/158 light client to fetch transactions for a Wallet.

Options:

List of Peer: Bitcoin full-nodes for the light client to connect to. May be empty.

connections: The number of connections for the light client to maintain.

scan_type: Sync, recover, or start a new wallet. For more information see ScanType.

data_dir: Optional directory to store block headers and peers.

A note on recovering wallets. Developers should allow users to provide an approximate recovery height and an estimated number of transactions for the wallet. When determining how many scripts to check filters for, the Wallet lookahead value will be used. To ensure all transactions are recovered, the lookahead should be roughly the number of transactions in the wallet history.

Implementations

impl CbfBuilder

pub fn new() -> Self

Start a new CbfBuilder

pub fn connections(&self, connections: u8) -> Arc<Self>

The number of connections for the light client to maintain. Default is two.

pub fn data_dir(&self, data_dir: String) -> Arc<Self>

Directory to store block headers and peers. If none is provided, the current working directory will be used.

pub fn scan_type(&self, scan_type: ScanType) -> Arc<Self>

Select between syncing, recovering, or scanning for new wallets.

pub fn peers(&self, peers: Vec<Peer>) -> Arc<Self>

Bitcoin full-nodes to attempt a connection with.

pub fn configure_timeout_millis( &self, handshake: u64, response: u64, ) -> Arc<Self>

Configure the time in milliseconds that a node has to:

Respond to the initial connection

Respond to a request

pub fn socks5_proxy(&self, proxy: Socks5Proxy) -> Arc<Self>

Configure connections to be established through a Socks5 proxy. The vast majority of the time, the connection is to a local Tor daemon, which is typically exposed at 127.0.0.1:9050`.

pub fn build(&self, wallet: &Wallet) -> CbfComponents

Construct a CbfComponents for a Wallet.