bdkffi::types
Enum WalletEvent

#[non_exhaustive]
pub enum WalletEvent {
    ChainTipChanged {
        old_tip: BlockId,
        new_tip: BlockId,
    },
    TxConfirmed {
        txid: Arc<Txid>,
        tx: Arc<Transaction>,
        block_time: ConfirmationBlockTime,
        old_block_time: Option<ConfirmationBlockTime>,
    },
    TxUnconfirmed {
        txid: Arc<Txid>,
        tx: Arc<Transaction>,
        old_block_time: Option<ConfirmationBlockTime>,
    },
    TxReplaced {
        txid: Arc<Txid>,
        tx: Arc<Transaction>,
        conflicts: Vec<Conflict>,
    },
    TxDropped {
        txid: Arc<Txid>,
        tx: Arc<Transaction>,
    },
}

Events representing changes to wallet transactions.

Returned after calling Wallet::apply_update_events.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ChainTipChanged

The latest chain tip known to the wallet changed.

Fields

old_tip: BlockId

Previous chain tip.

new_tip: BlockId

New chain tip.

TxConfirmed

A transaction is now confirmed.

If the transaction was previously unconfirmed old_block_time will be None.

If a confirmed transaction is now re-confirmed in a new block old_block_time will contain the block id and the time it was previously confirmed. This can happen after a chain reorg.

Fields

txid: Arc<Txid>

Transaction id.

tx: Arc<Transaction>

Transaction.

block_time: ConfirmationBlockTime

Confirmation block time.

old_block_time: Option<ConfirmationBlockTime>

Old confirmation block and time if previously confirmed in a different block.

TxUnconfirmed

A transaction is now unconfirmed.

If the transaction is first seen in the mempool old_block_time will be None.

If a previously confirmed transaction is now seen in the mempool old_block_time will contain the block id and the time it was previously confirmed. This can happen after a chain reorg.

Fields

txid: Arc<Txid>

Transaction id.

tx: Arc<Transaction>

Transaction.

old_block_time: Option<ConfirmationBlockTime>

Old confirmation block and time, if previously confirmed.

TxReplaced

An unconfirmed transaction was replaced.

This can happen after an RBF is broadcast or if a third party double spends an input of a received payment transaction before it is confirmed.

The conflicts field contains the txid and vin (in which it conflicts) of the conflicting transactions.

Fields

txid: Arc<Txid>

Transaction id.

tx: Arc<Transaction>

Transaction.

conflicts: Vec<Conflict>

Conflicting transactions.

TxDropped

Unconfirmed transaction dropped.

The transaction was dropped from the local mempool. This is generally due to the fee rate being too low. The transaction can still reappear in the mempool in the future resulting in a WalletEvent::TxUnconfirmed event.

Fields

txid: Arc<Txid>

Transaction id.

tx: Arc<Transaction>

Transaction.