bdkffi
Module kyoto

Structs

CbfBuilder

Build a BIP 157/158 light client to fetch transactions for a Wallet.

CbfClient

A CbfClient handles wallet updates from a CbfNode.

CbfComponents

Receive a CbfClient and CbfNode.

CbfNode

A CbfNode gathers transactions for a Wallet. To receive Update for Wallet, refer to the CbfClient. The CbfNode will run until instructed to stop.

IpAddress

An IP address to connect to over TCP.

Peer

A peer to connect to over the Bitcoin peer-to-peer network.

Socks5Proxy

A proxy to route network traffic, most likely through a Tor daemon. Normally this proxy is exposed at 127.0.0.1:9050.

Enums

Info

A log message from the node.

RecoveryPoint

ScanType

Sync a wallet from the last known block hash or recover a wallet from a specified recovery point.

Warning

Warnings a node may issue while running.