bdkffi
Module types

Structs

AddressInfo

A derived address and the index it was found at.

Anchor

Balance

Balance, differentiated into various categories.

BlockId

A reference to a block in the canonical chain.

CanonicalTx

A transaction that is deemed to be part of the canonical history.

ChainChange

The hash added or removed at the given height.

ChangeSet

Condition

An extra condition that must be satisfied but that is out of control of the user

ConfirmationBlockTime

Represents the confirmation block and time of a transaction.

Conflict

Represent a conflict in a replacement transaction.

EvictedTx

This type replaces the Rust tuple (txid, evicted_at) used in the Wallet::apply_evicted_txsmethod, whereevicted_atis the timestamp of when the transactiontxid` was evicted from the mempool. Transactions may be evicted for paying a low fee rate or having invalid scripts.

FullScanRequest

FullScanRequestBuilder

IndexerChangeSet

Mapping of descriptors to their last revealed index.

KeychainAndIndex

The keychain kind and the index in that keychain.

LocalChainChangeSet

Changes to the local chain

LocalOutput

An unspent output owned by a [Wallet].

MerkleProof

OutputStatus

Policy

Descriptor spending policy

ScriptAmount

A bitcoin script and associated amount.

SentAndReceivedValues

The total value sent and received.

SignOptions

Options for a software signer.

SyncRequest

SyncRequestBuilder

Tx

Bitcoin transaction metadata.

TxDetails

TxGraphChangeSet

TxStatus

Transaction confirmation metadata.

UnconfirmedTx

This type replaces the Rust tuple (tx, last_seen) used in the Wallet::apply_unconfirmed_txsmethod, wherelast_seenis the timestamp of when the transactiontx` was last seen in the mempool.

UniFfiTraitVtableFullScanScriptInspector

UniFfiTraitVtableSyncScriptInspector

Update

An update for a wallet containing chain, descriptor index, and transaction data.

Enums

ChainPosition

Represents the observed position of some chain data.

LockTime

PkOrF

Satisfaction

SatisfiableItem

WalletEvent

Events representing changes to wallet transactions.

Traits

FullScanScriptInspector

SyncScriptInspector

Functions

uniffi_bdkffi_fn_init_callback_vtable_fullscanscriptinspector

uniffi_bdkffi_fn_init_callback_vtable_syncscriptinspector